use long for the scale value (instead of int), as the scale values are potentially larger than an int, probably causing overflow exceptions
git-svn-id: https://josm.openstreetmap.de/osmsvn/applications/editors/josm/plugins@3614 b9d5c4c9-76e1-0310-9c85-f3177eceb1e4

+++ mappaint/src/mappaint/AreaElemStyle.java
@@ -5,7 +5,7 @@ public class AreaElemStyle extends ElemStyle
 {
 	Color colour;
 
-	public AreaElemStyle (Color colour, int maxScale, int minScale)
+	public AreaElemStyle (Color colour, long maxScale, long minScale)
 	{
 		this.colour = colour;
 		this.maxScale = maxScale;
