+++ mappaint/src/mappaint/IconElemStyle.java
@@ -6,7 +6,7 @@ public class IconElemStyle extends ElemStyle
 	ImageIcon icon;
 	boolean annotate;
 
-	public IconElemStyle (ImageIcon icon, boolean annotate, int maxScale, int minScale)
+	public IconElemStyle (ImageIcon icon, boolean annotate, long maxScale, long minScale)
 	{
 		this.icon=icon;
 		this.annotate=annotate;
